bootstrap: bitcheck: new utility to manipulate partition contents
Manipulating symlinks from a boot loader is not very convenient as
it requires support for the file system. Instead, there are many
unused bits in a partition table, so better create a new minimal
utility to manipulate and check such bits. It will then be easy to
modify the boot loader to check these bits to decide what image to
boot from.