bootstrap: scripts: introduce a new CLI for the firmware kernel
This script "firmware-cli" makes it possible with very minimal resources
to manage boot images and to download a new image over a serial port or
from the network. Since it's able to format the flash, it can be used to
complete the installation process once only the kernel and the rootfs
containing this script are put on the flash.