bootstrap: checkin of the tools necessary to build a bootloader.
All the unversionned work was imported. Kernel is updated to 2.6.16-rc4.
A busybox config file was added in order to ease later builds.

Right now, the makefile is limited to nsa1041, but the doc explains how
to build other ones too. The makefile only needs to get some parameters
for this.

The update-boot-image script from the rootfs was copied from the boot-scripts
repository. It will have to be maintained up to date and regularly picked from
the project.