bootstrap: mk-flash-layout : auto-detect the flash
Automatically detect where the flash is, possibly mount it, and fill in
/var/state/flash-layout.rc with appropriate information for other tools.
Right now it requires the software partition to be mountable (hda2), so
it is not completely autonomous yet for the boot loader. This will come
in a near future as this is less critical for most appliances.

This is an import from the boot-script repository, please don't develop
on this script here.