bootstrap: make firmware-cli pass "quiet" to the booted image
When firmware-cli boots the kernel itself, it can take ages because it
retrieves the parameters from a file which doesn't exist and never passes
the "quiet" argument, so kernel messages scroll on the 9600 Bps terminal.
Now it will automatically pass "quiet" if this keyword was present on the
running kernel's command line.