bootstrap: rootfs: fix busybox's rx to support an optional TTY name
By default, "rx" uses "/dev/tty". This doesn't work when the TTY is
mapped to /dev/console so in the past we've worked around this by
re-linking /dev/tty to /dev/ttyS0. But that breaks other utilities
like "more" which also need /dev/tty to read keystrokes.

Instead now we add an option to pass the tty to use to rx as a second
argument. A busybox patch is provided and serial-load was updated to
use it.